make frac second of legacy_hops_date double prec insteat of single

+++ source/cpp_src/Utilities/include/legacy_hops_date.hh
@@ -19,7 +19,8 @@ struct legacy_hops_date
         short day;
         short hour;
         short minute;
-        float second;
+        // float second;
+        double second;
 };
 
 } // namespace hops
